+++ aim/web/ui/src/components/Table/ArchiveModal.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import _ from 'lodash';
 import moment from 'moment';
 import classNames from 'classnames';
+import { react } from 'plotly.js';
 
 import Table from 'components/Table/Table';
 import { Button, Icon, Modal, Text } from 'components/kit';
@@ -12,7 +13,9 @@ function ArchiveModal({
   selectedRows,
   archiveMode,
   onRowSelect,
+  archiveRuns,
 }: any): React.FunctionComponentElement<React.ReactNode> {
+  let runsArchiveRequest: any = null;
   const [data, setData] = React.useState<any[]>([]);
   const [disabledData, setDisabledData] = React.useState<any[]>([]);
   const tableRef = React.useRef<any>({});
@@ -72,8 +75,14 @@ function ArchiveModal({
   ];
 
   React.useEffect(() => {
-    const archivedList: any[] = [];
-    const unarchivedList: any[] = [];
+    return () => {
+      runsArchiveRequest?.abort();
+    };
+  }, []);
+
+  React.useEffect(() => {
+    let archivedList: any[] = [];
+    let unarchivedList: any[] = [];
     const runHashList: string[] = [];
     Object.values(selectedRows || {}).forEach((selectedRow: any) => {
       if (!runHashList.includes(selectedRow.runHash)) {
@@ -81,12 +90,13 @@ function ArchiveModal({
         const rowData = {
           key: selectedRow.runHash,
           run: `${moment(selectedRow.creation_time * 1000).format(
-            'DD MMM YYYY, HH:mm A',
+            'DD MMM YYYY, HH:mm:ss A',
           )}`,
           experiment: selectedRow.experiment.name,
           runHash: selectedRow.runHash,
           selectKey: selectedRow.selectKey,
           isInProgress: !selectedRow?.end_time,
+          creationTime: selectedRow.creation_time * 1000000,
           isDisabled:
             (archiveMode && selectedRow.archived) ||
             (!archiveMode && !selectedRow.archived),
@@ -98,7 +108,8 @@ function ArchiveModal({
         }
       }
     });
-
+    archivedList = _.orderBy(archivedList, ['creationTime'], ['asc']);
+    unarchivedList = _.orderBy(unarchivedList, ['creationTime'], ['asc']);
     setData(archiveMode ? archivedList : unarchivedList);
     setDisabledData(!archiveMode ? archivedList : unarchivedList);
     tableRef.current?.updateData?.({
@@ -109,12 +120,18 @@ function ArchiveModal({
     });
   }, [selectedRows]);
 
+  function onArchive() {
+    const ids = data.map((item: any) => item.runHash);
+    runsArchiveRequest = archiveRuns(ids, archiveMode);
+    runsArchiveRequest.call().then(() => onClose());
+  }
+
   return (
     opened && (
       <Modal
         opened={opened}
         onClose={onClose}
-        onOk={() => {}}
+        onOk={onArchive}
         cancelButtonText='Cancel'
         okButtonText={archiveMode ? 'Archive' : 'Unarchive'}
         title={`Do you really want to ${
@@ -150,8 +167,10 @@ function ArchiveModal({
           )}
           {!_.isEmpty(disabledData) && (
             <Text size={12} weight={500} className='ActionModal__tableTitle'>
-              {`${Object.values(disabledData).length} Selected runs you can ${
-                !archiveMode ? 'archive' : 'unarchive'
+              {`${
+                Object.values(disabledData).length
+              } Selected runs you had already ${
+                archiveMode ? 'archived' : 'unarchived'
               }`}
             </Text>
           )}