+++ aim/web/ui/src/types/pages/scatters/Scatters.d.ts
@@ -60,6 +60,7 @@ export interface IScattersProps extends Partial<RouteChildrenProps> {
   requestIsPending: boolean;
   resizeMode: ResizeModeEnum;
   trendlineOptions: ITrendlineOptions;
+  selectedRows: any;
   onChangeTooltip: (tooltip: Partial<IChartTooltip>) => void;
   onChangeTrendlineOptions: (options: Partial<ITrendlineOptions>) => void;
   onActivePointChange?: (
@@ -99,4 +100,7 @@ export interface IScattersProps extends Partial<RouteChildrenProps> {
     delay?: number;
     enabled?: boolean;
   }) => void;
+  onRowSelect: any;
+  archiveRuns: (ids: string[], archived: boolean) => void;
+  deleteRuns: (ids: string[]) => void;
 }