+++ aim/web/ui/src/types/services/models/scatter/scatterAppModel.d.ts
@@ -13,9 +13,9 @@ import {
   IMetricsCollection,
 } from 'types/services/models/metrics/metricsAppModel';
 
-import { IMetricTrace, IRun } from './runModel';
-
 import { TrendlineTypeEnum } from 'utils/d3';
+
+import { IMetricTrace, IRun } from './runModel';
 
 export interface IScatterAppModelState {
   refs: {
@@ -32,6 +32,7 @@ export interface IScatterAppModelState {
   tableData: any[];
   tableColumns: ITableColumn[];
   sameValueColumns: string[];
+  selectedRows: any;
   params: string[];
   notifyData: INotification[];
   groupingSelectOptions: IGroupingSelectOption[];