+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - Add version, docs and slack links to sidebar (arsengit)
 - Enhance AudioPlayer component (arsengit)
 - Recover active tab in run details page after reload (roubkar)
-- Add ability to archive runs by batch (VkoHov)
+- Add ability to archive or delete runs batch (VkoHov)
 - Remote tracking server alpha version (alberttorosyan, mihran113, aramaim)
 - Add ability to change media elements order (VkoHov)
 - Add ability to hard delete runs (alberttorosyan)